Create hunter-lab10.html
My all caps runs but disappears immediately after click. I was so unsure how to find occurrences of a search character and how to add days to the date as well as substring for removal of the search character. I was able to get the word count. I have an issue with the logic part of writing the code.