Create hunter-lab11.html
This was difficult for me hard to grab the entered value to  sort search or reverse. I think I may have omitted the code for search. I tried also using the value of results after entering the array to sort or reverse and was unsuccessful. Maybe some videos of similar situations could be helpful.